front-service/requirements.txt
streamlit
streamlit-drawable-canvas
numpy
grpcio
grpcio-tools
